backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-dotenv
websockets
torch
torchaudio
chatterbox-tts
mlx-whisper
silero-vad
